Add edit complete test.js file
